Move generic transform into utils module.
